+++ src/_DataStructures_/Queue/index.js
@@ -1,19 +1,56 @@
-class Queue {
+const { LinkedList: SinglyLinkedLists, Node } = require('../LinkedList');
+
+class Queue extends SinglyLinkedLists {
   constructor() {
-    this.data = [];
+    super();
+    this.NotAllowed = 'Not Allowed';
   }
 
-  add(element) {
-    // add element to the start of the data
-    return this.data.unshift(element);
+  enqueue(data) {
+    const node = new Node(data, null);
+    return this.addAtEnd(node);
+  }
+
+  dequeue() {
+    const node = this.removeFromBeginning();
+    return node.data;
   }
 
   peek() {
-    return this.data[this.data.length - 1];
+    return this.getLast();
   }
 
-  remove() {
-    return this.data.pop();
+  size() {
+    return this.length();
+  }
+
+  destroy() {
+    this.delete();
+  }
+
+  /** Override and throw error for other LL methods */
+  addAtBeginning() {
+    throw new Error(this.NotAllowed);
+  }
+
+  addAt() {
+    throw new Error(this.NotAllowed);
+  }
+
+  removeFromEnd() {
+    throw new Error(this.NotAllowed);
+  }
+
+  getFirst() {
+    throw new Error(this.NotAllowed);
+  }
+
+  getAt() {
+    throw new Error(this.NotAllowed);
+  }
+
+  removeAt() {
+    throw new Error(this.NotAllowed);
   }
 }
